positioning now working
Changed to a 3DOF model with a 4th imaginary link of 0 length so that I can move the arm to any position without having to worry about the orientation of the 3rd link. Still need to make a function to know the end position of the end effector, partially done with angle calculations.